// src/generated/airline-alliance-lookup.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export interface paths {
    "/v1/lists/utilities/airlines/alliances": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["airlinesAlliancesLookup"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        AirlinesAlliancesLookupResponse: {
            /** @description Repeats associated airline alliance group information. */
            AllianceInfo?: {
                /**
                 * @description Returns the specified airline alliance group code(s).
                 * @example *O
                 */
                AllianceCode?: string;
                /**
                 * @description The name that corresponds to the specified airline alliance group code(s).
                 * @example oneworld
                 */
                AllianceName?: string;
                AirlineInfo?: {
                    /**
                     * @description The 2-character code of the alliance member airline.
                     * @example 4M
                     */
                    AirlineCode?: string;
                    Links?: Omit<components["schemas"]["CommonLinks"], "type">;
                }[];
            }[];
            Links?: Omit<components["schemas"]["CommonLinks"], "type">;
        };
        /** @description Array of links for different types of relations, e.g. self, linkTemplate, shop */
        CommonLinks: {
            /**
             * @description Describes relationship between href and current request.
             * @example self
             * @enum {string}
             */
            rel?: "self" | "shop" | "forecast" | "airportsInCity" | "airlineInfo" | "destinations" | "view" | "nextResults" | "prevResults" | "linkTemplate" | "shopTemplate" | "tagLookupLinkTemplate" | "paginatedRequestLinkTemplate";
            /**
             * @description Link to related API request.
             * @example https://api.cert.platform.sabre.com/v1/lists/utilities/airlines?airlinecode=AI
             */
            href?: string;
        }[];
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
    airlinesAlliancesLookup: {
        parameters: {
            query?: {
                /**
                 * @description 2-character airline alliance group code.
                 *
                 *     Multiple values are accepted.
                 *     Default: All airline alliance groups and their corresponding member information.
                 *     *A (Star Alliance™), *O (oneworld®), *S (SkyTeam).
                 * @example [
                 *       "*A"
                 *     ]
                 */
                alliancecode?: ("*A" | "*O" | "*S")[];
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description Success */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["AirlinesAlliancesLookupResponse"];
                };
            };
            /**
             * @description Bad Request
             *     - Parameter 'alliancecode' must be a valid value.
             */
            400: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
        };
    };
}